# add ascend npu accelerator support