--- a/src/transformers/testing_utils.py
+++ b/src/transformers/testing_utils.py
@@ -592,9 +592,7 @@ def require_torch_npu(test_case):
     """
     Decorator marking a test that requires NPU (in PyTorch).
     """
-    return unittest.skipUnless(is_torch_npu_available(), "test requires PyTorch NPU")(
-        test_case
-    )
+    return unittest.skipUnless(is_torch_npu_available(), "test requires PyTorch NPU")(test_case)
 
 
 def require_torch_multi_npu(test_case):