@HuggingFaceDocBuilderDev 2023-07-18
The docs for this PR live [here](https://moon-ci-docs.huggingface.co/docs/transformers/pr_24879). All of your documentation changes will be reflected on that endpoint.